+++ src/bpp/models/patent.py
@@ -1,6 +1,9 @@
 # -*- encoding: utf-8 -*-
+from denorm import denormalized, depend_on_related
 from django.db import models
 from django.db.models import CASCADE, SET_NULL
+
+from django.contrib.postgres.fields import ArrayField, JSONField
 
 from django.utils.functional import cached_property
 
@@ -13,6 +16,7 @@ from bpp.models import (
     ModelZeStatusem,
     ModelZeSzczegolami,
     ModelZInformacjaZ,
+    ModelZPrzeliczaniemDyscyplin,
     ModelZRokiem,
     ModelZWWW,
 )
@@ -71,6 +75,7 @@ class Patent(
     DodajAutoraMixin,
     ModelZAbsolutnymUrl,
     ModelOpcjonalnieNieEksportowanyDoAPI,
+    ModelZPrzeliczaniemDyscyplin,
 ):
     tytul_oryginalny = models.TextField("Tytuł oryginalny", db_index=True)
 
@@ -117,3 +122,77 @@ class Patent(
 
     def clean(self):
         self.tytul_oryginalny = safe_html(self.tytul_oryginalny)
+
+    #
+    # Cache framework by django-denorm-iplweb
+    #
+
+    denorm_always_skip = ("ostatnio_zmieniony",)
+
+    @denormalized(JSONField, blank=True, null=True)
+    @depend_on_related(
+        "bpp.Patent_Autor",
+        only=(
+            "typ_odpowiedzialnosci_id",
+            "afiliuje",
+            "dyscyplina_naukowa_id",
+            "upowaznienie_pbn",
+            "przypieta",
+        ),
+    )
+    def cached_punkty_dyscyplin(self):
+        # TODO: idealnie byłoby uzależnić zmiane od pola 'rok' które by było identyczne
+        # dla bpp.Poziom_Wydawcy, rok i id z nadrzędnego. Składnia SQLowa ewentualnie
+        # jakis zapis django-podobny mile widziany.
+        return self.przelicz_punkty_dyscyplin()
+
+    @denormalized(models.TextField, default="")
+    @depend_on_related(
+        "bpp.Patent_Autor",
+        only=("zapisany_jako", "typ_odpowiedzialnosci_id", "kolejnosc"),
+    )
+    @depend_on_related("bpp.Status_Korekty")
+    def opis_bibliograficzny_cache(self):
+        return self.opis_bibliograficzny()
+
+    @denormalized(ArrayField, base_field=models.TextField(), blank=True, null=True)
+    @depend_on_related(
+        "bpp.Autor",
+        only=(
+            "nazwisko",
+            "imiona",
+        ),
+    )
+    @depend_on_related("bpp.Patent_Autor", only=("kolejnosc",))
+    def opis_bibliograficzny_autorzy_cache(self):
+        return [
+            "%s %s" % (x.autor.nazwisko, x.autor.imiona)
+            for x in self.autorzy_dla_opisu()
+        ]
+
+    @denormalized(models.TextField, blank=True, null=True)
+    @depend_on_related(
+        "bpp.Patent_Autor",
+        only=("zapisany_jako", "kolejnosc"),
+    )
+    def opis_bibliograficzny_zapisani_autorzy_cache(self):
+        return ", ".join([x.zapisany_jako for x in self.autorzy_dla_opisu()])
+
+    @denormalized(
+        models.SlugField,
+        max_length=400,
+        unique=True,
+        db_index=True,
+        null=True,
+        blank=True,
+    )
+    @depend_on_related(
+        "bpp.Patent_Autor",
+        only=("zapisany_jako", "kolejnosc"),
+    )
+    @depend_on_related(
+        "bpp.Autor",
+        only=("nazwisko", "imiona"),
+    )
+    def slug(self):
+        return self.get_slug()