+++ src/bpp/models/wydawnictwo_zwarte.py
@@ -2,12 +2,14 @@
 import re
 import warnings
 
+from denorm import denormalized, depend_on_related
 from dirtyfields.dirtyfields import DirtyFieldsMixin
 from django.db import models
 from django.db.models import CASCADE, PROTECT
 from django.db.models.expressions import RawSQL
 
 from django.contrib.contenttypes.fields import GenericRelation
+from django.contrib.postgres.fields import ArrayField, JSONField
 
 from bpp.models import (
     DodajAutoraMixin,
@@ -36,6 +38,7 @@ from bpp.models.abstract import (
     ModelZKonferencja,
     ModelZLiczbaCytowan,
     ModelZOpenAccess,
+    ModelZPrzeliczaniemDyscyplin,
     ModelZPubmedID,
     ModelZRokiem,
     ModelZSeria_Wydawnicza,
@@ -176,6 +179,7 @@ class Wydawnictwo_Zwarte(
     MaProcentyMixin,
     DodajAutoraMixin,
     DirtyFieldsMixin,
+    ModelZPrzeliczaniemDyscyplin,
 ):
     """Wydawnictwo zwarte, czyli: książki, broszury, skrypty, fragmenty,
     doniesienia zjazdowe."""
@@ -232,6 +236,86 @@ class Wydawnictwo_Zwarte(
                 "",
             )
         )
+
+    #
+    # Cache framework by django-denorm-iplweb
+    #
+
+    denorm_always_skip = ("ostatnio_zmieniony",)
+
+    @denormalized(JSONField, blank=True, null=True)
+    @depend_on_related(
+        "bpp.Wydawnictwo_Zwarte_Autor",
+        only=(
+            "typ_odpowiedzialnosci_id",
+            "afiliuje",
+            "dyscyplina_naukowa_id",
+            "upowaznienie_pbn",
+            "przypieta",
+        ),
+    )
+    @depend_on_related("bpp.Wydawca", only=("lista_poziomow", "alias_dla_id"))
+    def cached_punkty_dyscyplin(self):
+        # TODO: idealnie byłoby uzależnić zmiane od pola 'rok' które by było identyczne
+        # dla bpp.Poziom_Wydawcy, rok i id z nadrzędnego. Składnia SQLowa ewentualnie
+        # jakis zapis django-podobny mile widziany.
+        return self.przelicz_punkty_dyscyplin()
+
+    @denormalized(models.TextField, default="")
+    @depend_on_related("self", "wydawnictwo_nadrzedne")
+    @depend_on_related(
+        "bpp.Wydawnictwo_Zwarte_Autor",
+        only=("zapisany_jako", "typ_odpowiedzialnosci_id", "kolejnosc"),
+    )
+    @depend_on_related("bpp.Wydawca", only=("nazwa", "alias_dla_id"))
+    @depend_on_related("bpp.Charakter_Formalny")
+    @depend_on_related("bpp.Typ_KBN")
+    @depend_on_related("bpp.Status_Korekty")
+    def opis_bibliograficzny_cache(self):
+        return self.opis_bibliograficzny()
+
+    @denormalized(ArrayField, base_field=models.TextField(), blank=True, null=True)
+    @depend_on_related(
+        "bpp.Autor",
+        only=(
+            "nazwisko",
+            "imiona",
+        ),
+    )
+    @depend_on_related("bpp.Wydawnictwo_Zwarte_Autor", only=("kolejnosc",))
+    def opis_bibliograficzny_autorzy_cache(self):
+        return [
+            "%s %s" % (x.autor.nazwisko, x.autor.imiona)
+            for x in self.autorzy_dla_opisu()
+        ]
+
+    @denormalized(models.TextField, blank=True, null=True)
+    @depend_on_related(
+        "bpp.Wydawnictwo_Zwarte_Autor",
+        only=("zapisany_jako", "kolejnosc"),
+    )
+    def opis_bibliograficzny_zapisani_autorzy_cache(self):
+        return ", ".join([x.zapisany_jako for x in self.autorzy_dla_opisu()])
+
+    @denormalized(
+        models.SlugField,
+        max_length=400,
+        unique=True,
+        db_index=True,
+        null=True,
+        blank=True,
+    )
+    @depend_on_related(
+        "bpp.Wydawnictwo_Zwarte_Autor",
+        only=("zapisany_jako", "kolejnosc"),
+    )
+    @depend_on_related(
+        "bpp.Autor",
+        only=("nazwisko", "imiona"),
+    )
+    @depend_on_related("self", "wydawnictwo_nadrzedne")
+    def slug(self):
+        return self.get_slug()
 
 
 class Wydawnictwo_Zwarte_Zewnetrzna_Baza_Danych(models.Model):