+++ src/django_bpp/settings/base.py
@@ -145,6 +145,7 @@ if TESTING:
 
 
 INSTALLED_APPS = [
+    "denorm.apps.DenormAppConfig",
     "reversion",
     "djangoql",
     "cacheops",
@@ -671,15 +672,5 @@ PERMISSIONS_WIDGET_EXCLUDE_MODELS = [
     "bpp.cachequeue",
 ]
 
-
-CACHEOPS = {
-    "bpp.bppmultiseekvisibility": {"ops": ("get", "fetch")},
-    "dbtemplates.template": {"ops": ("fetch", "get")},
-    "bpp.szablondlaopisubibliograficznego": {"ops": ("fetch", "get")},
-    "miniblog.article": {"ops": ("get", "fetch")},
-}
-CACHEOPS_REDIS = BROKER_URL
-
-CACHEOPS_DEFAULTS = {"timeout": 60 * 60}
 
 DBTEMPLATES_USE_REVERSION = True